requirements.txt
absl-py==2.3.1​
cachetools==5.5.2​
google-auth==2.41.1​
google-auth-oauthlib==1.0.0​
grpcio==1.70.0​
importlib-metadata==8.5.0​
lmdb==1.7.3​
markdown==3.7​
oauthlib==3.3.1​
protobuf==5.29.5​
pyasn1==0.6.1​
pyasn1-modules==0.4.2​
requests==2.32.4​
requests-oauthlib==2.0.0​
rsa==4.9.1​
six==1.17.0​
tensorboard==2.14.0​
tensorboard-data-server==0.7.2​
werkzeug==3.0.6​
zipp==3.20.2​
pyyaml​
sympy​
networkx​
pillow​
numpy​
urllib3​
charset-normalizer​
idna​
filelock​
jinja2​
markupsafe